Change to start/end date variables creation - reduce variables needed and change names
Reduce the number of variables created to create the individual start date (index date) and end date used in the inclusion/exclusion criteria.
And rename them.
End_date is commented out has it doesn't seem to be used in the code to check criteria, this later point needs to be double checked.