Include command arguments to read entry dataset/parameter and if/else in part 3a&3c
In project.yaml, the dataset (and maybe another argument) will be specified as entry for the action to run this stage 1 R code. There will be 2 actions to run, one for the vaccinated cohort and one for the electively unvaccinated cohort.
The entry argument(s) whether for the vaccinated cohort or electively unvaccinated cohort will determine which part of the code will be used for part 3a (index_start_date) and for part 3c (criteria 7-11 for vax, a different criteria 7 for e_unvax).
The code will be further updated once the two datasets for vaccinated and electively unvaccinated subcohorts will have been created from a new study definition code.