Remove QA and incl/excl variables from stage1 datasets
In stage 1 R code, remove QA variables and the variables created to apply inclusion/exclusion criteria from the final stage 1 datasets.
This is done to reduce the size of the dataset for efficient reason when the analysis is running on real data.
Keep the name of the vax variables consistent i.e. using `vax_` rather than `vacc_`